2004 version of the site
This is a back-dated commit of what likely was one of the first "proper"
versions of this site (I don't recall exactly). This is not an exact
historically accurate copy, but it's probably close to the original.
However, several files included here are in a more recent version rather
than the original. In most of these cases, the difference is minor. The
reason for including more recent versions is either that the originals
no longer exist or that it isn't entirely clear if the older versions
would give an accurate representation when deployed in a modern
environment.