2006 version of the site
The "proper" PHP version of 2004 wasn't maintained for very long. This
is a back-dated commit of its 2006 replacement. Parts of the older
versions actually remained online for a long time. This commit is not
an exact historically accurate copy and does not reflect that.

The "professional" home page almost certainly already existed by 2006.
However, its repository was only started in late 2007 for the first
tutorials.