2012 version of the site
The main site is replaced by something that at least is somewhat
pleasant to look at. This back-dated commit contains a more or less
historically accurate copy of the version as it was at the time. Also,
the "professional" home page's repository is added as a submodule, but
the repository was not in fact synced with what was actually online at
the time.