Fix invisible text in Safari
Safari appears to have an issue causing a background image on the <html>
element to cover all content on the page when it has a z-index. This
style sheet triggered that bug (with a z-index of -1, which should if
anything push the background image further back, not to the front).
Chrome used to suffer, too, up to at least version 50, at least
according to the Chromium issue tracker <https://bugs.chromium.org/p/chromium/issues/detail?id=367703>,
but that is no longer reproducible with the most recent Chrome version.
iCab, which uses WebKit as well, strangely never exhibited this issue.