New home page design: Replace prose with images.
This is a first pass only. I've taken a number of shortcuts with this page design in order to rush it out the door.

One of the issues is the lack of fully semantic markup. For example, the link lists in the footer should have one list item per link; instead, the list items contain groups of links separated by line breaks. This'll do for now, even though it is not really good enough. What we'd need here is a list of lists of links (with the inner lists representing the groups), but that will have to come later.

Furthermore, the contact details are perhaps better marked up as a definition list…? Need to recheck the spec here.

Another area of concern is the selection and order of the portfolio images. This should perhaps be automated rather than hand-crafted hard-coded HTML. But this is both less important and less urgent than HTML semantics I guess.

Semantically, however, the individual portfolio items probably aren't sections.

The actual portfolio images are not included in this commit.

Finally, the CSS should be optimised. In particular, I'm concerned about compatibility of calc(), and about graceful degradation in general.